v2.0.4
All functionalities of server are restored
