+++ src/main/java/demo/geo/app/controllers/ExportXlsController.java
@@ -1,50 +1,36 @@
-package demo.geo.app.xls;
+package demo.geo.app.controllers;
 
-import io.swagger.annotations.ApiOperation;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.Resource;
 import org.springframework.core.io.FileSystemResource;
+import io.swagger.annotations.ApiOperation;
 
 import java.net.MalformedURLException;
 import java.io.File;
 import java.io.IOException;
 import javax.servlet.http.HttpServletResponse;
 
-import demo.geo.app.entities.Section;
-import demo.geo.app.exceptions.NotFoundException;
-import demo.geo.app.model.SectionRepository;
 import demo.geo.app.xls.enums.JStatus;
 import demo.geo.app.xls.enums.JType;
+import demo.geo.app.entities.Job;
+import demo.geo.app.services.XlsService;
 
 @RestController
 @RequestMapping("/export")
 public class ExportXlsController {
 
-    private final SectionRepository sectionRepository;
-
     private final XlsService xmlService;
-    
-    private final FileService fileService;
-    
+        
     @Autowired
-    public ExportXlsController(SectionRepository sectionRepository, 
-            XlsService xmlService, FileService fileService) {
-        this.sectionRepository = sectionRepository;
+    public ExportXlsController(XlsService xmlService) {
         this.xmlService = xmlService;
-        this.fileService = fileService;
     }
     
     @GetMapping("")
-    @ApiOperation("Returns ID of the Async Job and launches exporting")
-    public String exportSections() {
-        Iterable<Section> sections = sectionRepository.findAll();
-        if (sections.toString().equals("[]")) {
-            throw new NotFoundException("! No any section found (DB is empty) !");
-        }
-        Job job = xmlService.startJob(JType.EXPORT);
-        fileService.generateXLS(job);
-        return "{ \"Job ID\" : " + job.getId().toString() + "}";
+    @ApiOperation("Returns Async Job and launches exporting DB to the xls file")
+    public Job exportSections() {
+        return xmlService.startJob(JType.EXPORT);
     }
     
     @GetMapping("/{id}")
