+++ src/main/java/demo/geo/app/services/FileService.java
@@ -1,6 +1,6 @@
-package demo.geo.app.xls;
+package demo.geo.app.services;
 
-import org.springframework.web.multipart.MultipartFile;
+import demo.geo.app.entities.Job;
 import org.springframework.stereotype.Service;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.annotation.Async;
@@ -8,23 +8,18 @@ import org.apache.poi.hssf.usermodel.*;
 
 import java.util.*;
 import java.io.IOException;
-import java.io.InputStream;
-import java.net.MalformedURLException;
-import java.time.LocalDateTime;
 import java.util.List;
-import java.io.BufferedOutputStream;
 import java.io.File;
 import java.io.FileOutputStream;
-import java.io.FileNotFoundException;
 
 import demo.geo.app.entities.GeologicalClass;
 import demo.geo.app.entities.Section;
-import demo.geo.app.exceptions.NotFoundException;
-import demo.geo.app.exceptions.UnprocException;
-import demo.geo.app.model.SectionRepository;
+import demo.geo.app.dao.SectionRepository;
+import demo.geo.app.dao.JobRepository;
 import demo.geo.app.xls.enums.JStatus;
-import demo.geo.app.xls.enums.JType;
+import java.io.FileInputStream;
 
+@Service
 public class FileService {
     
     private final SectionRepository sectionRepository;
@@ -94,6 +89,66 @@ public class FileService {
         HSSFCell newCell = row.createCell(Math.max(row.getLastCellNum(), 0));
         newCell.setCellValue(stringValue);
     }
+
+    /**
+     * Reads uploaded file and parses it as XLS book; 
+     * Also sets up the job status 
+     * (for TZ#3: /import)
+     * 
+     * @param file file to parse
+     * @param job current job of importing
+     */
+    @Async
+    public void parseXLS(File file, Job job) {
+        try {
+            FileInputStream inputStream = new FileInputStream(file);
+            HSSFWorkbook xlsFile = new HSSFWorkbook(inputStream);
+            inputStream.close();
+            HSSFSheet sheet = xlsFile.getSheetAt(0); // header
+            // Read body
+            for (int i = 1; i < sheet.getPhysicalNumberOfRows(); i++) {
+                // Read string
+                HSSFRow currentRow = sheet.getRow(i);
+                if (currentRow == null) {
+                    continue;
+                }
+                String secName = currentRow.getCell(0).getStringCellValue();
+                Section section = sectionRepository.save(new Section(secName));
+
+                // Jumping for pairs {geoClass, geoCode}
+                List<GeologicalClass> listOfGeoClasses = new ArrayList<>();
+                for (int j = 1; j < currentRow.getLastCellNum(); j += 2) {
+                    String geoClassName = readCell(currentRow, j);
+                    String geoClassCode = readCell(currentRow, j+1);
+                    if (geoClassName == null || geoClassCode == null) {
+                        continue;
+                    }
+                    listOfGeoClasses.add(new GeologicalClass(section.getId(), geoClassName, geoClassCode));
+                }
+                section.addListOfGeoClasses(listOfGeoClasses);
+                sectionRepository.save(section);
+            }
+            job.setStatus(JStatus.DONE);
+            jobRepository.save(job);
+        } catch (IOException e) {
+            job.setStatus(JStatus.ERROR);
+            jobRepository.save(job);
+        }
+    }
     
+    /**
+     * Auxiliary func for reading geoClassName or geoClassCode from HSSF cell
+     * 
+     * @param currentRow current row to read cell
+     * @param cellNum number of cell in current row to read 
+     * @return readed text value 
+     */
+    private String readCell (HSSFRow currentRow, int cellNum) {
+        HSSFCell geoNameCell = currentRow.getCell(cellNum);
+        if(geoNameCell == null) {
+            return null;
+        }
+        return geoNameCell.getStringCellValue();
+    }
     
 }
